move date validation steps
