+++ app/routes/enrich-request.js
@@ -1,11 +1,10 @@
 const { getPaymentRequestByInvoiceNumberAndRequestId } = require('../payment-request')
-const format = require('../utils/date-formatter')
 const ViewModel = require('./models/enrich-request')
 const enrichRequestSchema = require('./schemas/enrich-request')
-const dateSchema = require('./schemas/date')
 const { enrichment } = require('../auth/permissions')
 const { getUser } = require('../auth')
 const { enrichRequest } = require('../processing/enrich')
+const { validateDate } = require('../processing/validate-date')
 
 module.exports = [{
   method: 'GET',
@@ -53,30 +52,9 @@ module.exports = [{
         return h.view('enrich-request', { paymentRequest, ...new ViewModel(payload, enrichRequestValidation.error) }).code(400).takeover()
       }
 
-      const day = format(payload.day)
-      const month = format(payload.month)
-      const year = payload.year
-
-      const validDate = dateSchema({
-        date: `${year}-${month}-${day}`
-      }, paymentRequest.received)
-
-      if (validDate.error) {
-        return h.view('enrich-request', { paymentRequest, ...new ViewModel(payload, validDate.error) }).code(400).takeover()
-      }
-
-      if (payload?.day === '29' && payload?.month === '02') {
-        const isLeap = new Date(payload?.year, 1, 29).getDate() === 29
-        if (!isLeap) {
-          const leapError = {
-            details: [{
-              context: {
-                label: 'date-not-leap-year'
-              }
-            }]
-          }
-          return h.view('enrich-request', { paymentRequest, ...new ViewModel(payload, leapError) }).code(400).takeover()
-        }
+      const dateError = await validateDate(payload, paymentRequest.received)
+      if (dateError) {
+        return h.view('enrich-request', { paymentRequest, ...new ViewModel(payload, dateError) }).code(400).takeover()
       }
 
       if (paymentRequest.released) {
